requirements.txt
fastapi
uvicorn
aiohttp
aiofiles
opencv-python
torch
torchvision
detectron2
tesserocr
face_liveness
prnet @ git+https://github.com/YadiraF/PRNet.git
deca
open3d-python
pycpd
faiss-cpu
# Pin NumPy <2 to satisfy scikit-learn/pandas/mediapipe wheels
numpy==1.26.4
scipy
meshio
trimesh
pygltflib
Pillow
# OSINT dependencies
playwright
face_recognition
beautifulsoup4
feedparser
newspaper3k
requests
selenium
googlesearch-python
python-whois
shodan
phonenumbers
social-analyzer
# Enhanced Facial Pipeline Dependencies
mediapipe>=0.10.0
dlib>=19.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
